feat: refatorando tela backlog para mvc
